+++ app/Http/Controllers/ProfileController.php
@@ -6,7 +6,9 @@ use App\Http\Requests\ProfileUpdateRequest;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Redirect;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\View\View;
 
 class ProfileController extends Controller
@@ -26,7 +28,20 @@ class ProfileController extends Controller
      */
     public function update(ProfileUpdateRequest $request): RedirectResponse
     {
-        $request->user()->fill($request->validated());
+        if ($request->file('photo')) {
+            $filename = basename($request->file('photo')->getClientOriginalName(), '.' . $request->file('photo')->getClientOriginalExtension());
+            $path = $request->file('photo')->storeAs('/assets/profile-user', $filename . "-" . Auth::user()->email . "." . $request->file('photo')->getClientOriginalExtension(), 'public');
+            $data = $request->validated();
+
+            $data['photo'] = $path;
+
+            if (!is_null($request->user()->photo)) {
+                Storage::disk('public')->delete($request->user()->photo);
+            }
+            $request->user()->fill($data);
+        } else {
+            $request->user()->fill($request->validated());
+        }
 
         if ($request->user()->isDirty('email')) {
             $request->user()->email_verified_at = null;