+++ app/Models/ContentCreatorSkill.php
@@ -4,8 +4,16 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class ContentCreatorSkill extends Model
 {
     use HasFactory;
+
+    protected $guarded = ['id'];
+
+    public function creator(): BelongsTo
+    {
+        return $this->belongsTo(ContentCreator::class, 'creator_id');
+    }
 }