+++ database/migrations/2023_04_06_151621_create_content_creators_table.php
@@ -14,12 +14,10 @@ return new class extends Migration
         Schema::create('content_creators', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('category_id');
             $table->text('description')->nullable();
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users');
-            $table->foreign('category_id')->references('id')->on('content_creator_categories');
         });
     }