+++ resources/views/landing-page.blade.php
@@ -1,0 +1,175 @@
+<x-main-layout>
+    <div class="hero">
+        <div class="hero-content flex-col lg:flex-row">
+            <img src="assets/images/slider/slider1.png" class=" rounded-lg shadow-sm sm:max-w-sm lg:max-w-lg" />
+            <div class="flex flex-col justify-center items-center">
+                <h1 class="text-5xl font-bold text-center sm:text-7xl sm:text-left">jadiUMKM</h1>
+                <p class="py-6 font-bold text-2xl">Learn. Grow. Thrive.
+                </p>
+                <a class="btn border-none hover:text-white sm:content-center" href="{{ route('register') }}">Get
+                    Started</a>
+            </div>
+        </div>
+    </div>
+
+    <div class="divider"></div>
+
+    <div class="grid grid-cols-1 justify-center gap-4 sm:grid-cols-3 text-left">
+        <div class="card shadow-xl bg-unyu ">
+            <div class="card-body ">
+                <h2 class="card-title">Apa Manfaat Digitalisasi UMKM?</h2>
+                <p class="pt-3 font-bold">Memperluas jangkauan</p>
+                <p>
+                    Digitalisasi UMKM memungkinkan pelaku usaha memiliki jangkauan pemasaran yang lebih luas bahkan bisa
+                    sampai ke pelosok negeri.</p>
+                <p class="pt-3 font-bold">Memperluas jangkauan</p>
+                <p>Meningkatkan pendapatan
+                    Pastinya, semakin luas jangkauan pemasaran suatu bisnis maka penghasilan pun akan meningkat.</p>
+                <p class="pt-3 font-bold">Memperluas jangkauan</p>
+                <p>
+                    Mempermudah transaksi
+                    Pembayaran digital akan memudahkan UMKM dan konsumen. Goodbye kembalian!</p>
+                <p class="pt-3 font-bold">Memperluas jangkauan</p>
+                <p>Mengikuti trend
+                    Digitalisasi membuat UMKM beradaptasi dengan kebiasaan masyarakat sehingga lebih kekinian.</p>
+            </div>
+        </div>
+        <div class="card  shadow-xl bg-unyu">
+            <div class="card-body ">
+                <h2 class="card-title">Bagaimana Strategi Untuk Mengembangkan Usaha Kecil Yang Dijalankan?</h2>
+                <p class="pt-4">Berdasarkan informasi yang di dapatkan, ternyata sekitar 70% UKM yang ada di Indonesia
+                    memulai UKM tersebut karena adanya desakan ekonomi bukan karena mereka memiliki produk yang unik
+                    atau keterampilan pada bidang tertentu. Tentu saja kondisi ini akhirnya membuat sebagian besar dari
+                    UKM di Indonesia tidak memiliki daya saing, dimana kita ketahui bahwa untuk tetap bertahan dan
+                    berkembang di dalam dunia bisnis yang semakin ketat kita harus memiliki keterampilan, dapat bekerja
+                    secara profesional, dan mampu menciptakan inovasi-inovasi pada bisnis mereka.</p>
+            </div>
+        </div>
+        <div class="card  shadow-xl bg-unyu">
+            <div class="card-body">
+                <h2 class="card-title">Bagaimana Cara Membuat Konten Digital dengan Nilai Edukatif dan Menghibur secara
+                    Konsisten?</h2>
+                <p class="pt-4">Nilai utama dalam dunia digital adalah kreativitas, kolaborasi, dan kritis. Dengan
+                    tiga kemampuan ini, pengguna media digital dapat berkontribusi pada berbagai hal positif, seperti
+                    meningkatkan ide dan solusi di ranah digital.</p>
+            </div>
+        </div>
+    </div>
+
+
+    <div class="divider"></div>
+
+    <div class="relative isolate overflow-hidden bg-white py-16 sm:py-32">
+        <div class="hidden sm:absolute sm:-top-10 sm:right-1/2 sm:-z-10 sm:mr-10 sm:block sm:transform-gpu sm:blur-3xl"
+            aria-hidden="true">
+            <div class="aspect-[1097/845] w-[68.5625rem] bg-gradient-to-tr from-[#ff4694] to-[#776fff] opacity-20"
+                style="clip-path: polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)">
+            </div>
+        </div>
+        <div class="absolute -top-52 left-1/2 -z-10 -translate-x-1/2 transform-gpu blur-3xl sm:top-[-28rem] sm:ml-16 sm:translate-x-0 sm:transform-gpu"
+            aria-hidden="true">
+            <div class="aspect-[1097/845] w-[68.5625rem] bg-gradient-to-tr from-[#ff4694] to-[#776fff] opacity-20"
+                style="clip-path: polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)">
+            </div>
+        </div>
+        <div class="mx-auto max-w-7xl px-6 lg:px-8">
+            <div class="mx-auto max-w-2xl lg:mx-0">
+                <h2 class="text-4xl font-bold tracking-tight text-black sm:text-6xl">About Us</h2>
+                <p class="mt-6 text-lg leading-8 text-black">JadiUMKM adalah platform pembelajaran terkait bisnis dan
+                    digitalisasi untuk mengembangkan bisnis Anda. Kami memberi fitur dan pengetahuan yang Anda butuhkan
+                    untuk berkembang — baik untuk peluncuran produk baru atau upaya rebranding, kami ingin membantu Anda
+                    mengembangkan bisnis melalui berbagi pengetahuan, rekomendasi, dan lainnya.
+                </p>
+                <p class="mt-6 text-lg leading-8 text-black">Kami percaya bahwa UMKM adalah salah satu kunci pertumbuhan
+                    ekonomi negara kita, kami ingin membantu Anda dalam belajar dan mengembangkan bisnis!
+                </p>
+            </div>
+            <div class="mx-auto mt-10 max-w-2xl lg:mx-0 lg:max-w-none">
+                <div
+                    class="grid grid-cols-1 gap-x-8 gap-y-6 text-base font-semibold leading-7 text-black sm:grid-cols-2 md:flex lg:gap-x-10">
+                    <a href="{{ route('video') }}">Pembelajaran <span aria-hidden="true">&rarr;</span></a>
+
+                    <a href="https://t.me/+BielMi-IEV5mOGQ1" target="_blank">Pembelajaran Premium<span
+                            aria-hidden="true">&rarr;</span></a>
+
+                    <a href="{{ route('webinar') }}">Webinar <span aria-hidden="true">&rarr;</span></a>
+
+                    <a href="{{ route('creator') }}">Content Creator <span aria-hidden="true">&rarr;</span></a>
+
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <div class="divider"></div>
+
+    <div class="bg-unyu px-6 py-16 sm:py-16 lg:px-8">
+        <div class="absolute inset-x-0 top-[-10rem] -z-10 transform-gpu overflow-hidden blur-3xl sm:top-[-20rem]"
+            aria-hidden="true">
+            <div class="relative left-1/2 -z-10 aspect-[1155/678] w-[36.125rem] max-w-none -translate-x-1/2 rotate-[30deg] bg-gradient-to-tr from-[#ff80b5] to-[#9089fc] opacity-30 sm:left-[calc(50%-40rem)] sm:w-[72.1875rem]"
+                style="clip-path: polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)">
+            </div>
+        </div>
+        <div class="mx-auto max-w-2xl text-center">
+            <h2 class="text-3xl font-bold tracking-tight text-gray-900 sm:text-5xl">Contact Us</h2>
+        </div>
+        <form action="https://formspree.io/f/mgebpkwk" method="POST" class="mx-auto mt-16 max-w-xl sm:mt-20">
+            <div class="grid grid-cols-1 gap-x-8 gap-y-6 sm:grid-cols-2">
+                <div>
+                    <label for="first-name" class="block text-sm font-semibold leading-6 text-gray-900">First
+                        name</label>
+                    <div class="mt-2.5">
+                        <input type="text" name="first-name" id="first-name" autocomplete="given-name"
+                            class="block w-full rounded-md border-0 px-3.5 py-2 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6">
+                    </div>
+                </div>
+                <div>
+                    <label for="last-name" class="block text-sm font-semibold leading-6 text-gray-900">Last name</label>
+                    <div class="mt-2.5">
+                        <input type="text" name="last-name" id="last-name" autocomplete="family-name"
+                            class="block w-full rounded-md border-0 px-3.5 py-2 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6">
+                    </div>
+                </div>
+                <div class="sm:col-span-2">
+                    <label for="company" class="block text-sm font-semibold leading-6 text-gray-900">Company</label>
+                    <div class="mt-2.5">
+                        <input type="text" name="company" id="company" autocomplete="organization"
+                            class="block w-full rounded-md border-0 px-3.5 py-2 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6">
+                    </div>
+                </div>
+                <div class="sm:col-span-2">
+                    <label for="email" class="block text-sm font-semibold leading-6 text-gray-900">Email</label>
+                    <div class="mt-2.5">
+                        <input type="email" name="email" id="email" autocomplete="email"
+                            class="block w-full rounded-md border-0 px-3.5 py-2 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6">
+                    </div>
+                </div>
+                <div class="sm:col-span-2">
+                    <label for="phone-number" class="block text-sm font-semibold leading-6 text-gray-900">Phone
+                        number</label>
+                    <div class="relative mt-2.5">
+                        <input type="tel" name="phone-number" id="phone-number" autocomplete="tel"
+                            class="block w-full rounded-md border-0 px-3.5 py-2 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6">
+                    </div>
+                </div>
+                <div class="sm:col-span-2">
+                    <label for="message" class="block text-sm font-semibold leading-6 text-gray-900">Message</label>
+                    <div class="mt-2.5">
+                        <textarea name="message" id="message" rows="4"
+                            class="block w-full rounded-md border-0 px-3.5 py-2 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"></textarea>
+                    </div>
+                </div>
+            </div>
+            <div class="mt-10 text-white">
+                <button type="submit"
+                    class="btn border-none block w-full rounded-md px-3.5 py-2.5 text-center text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600">Let's
+                    talk</button>
+            </div>
+        </form>
+    </div>
+
+    <div class="divider"></div>
+
+
+
+</x-main-layout>