+++ routes/web.php
@@ -1,6 +1,15 @@
 <?php
 
+use App\Http\Controllers\Admin\Edukasi\Video\VideoController;
+use App\Http\Controllers\Admin\EdukasiController;
 use App\Http\Controllers\ProfileController;
+use App\Http\Controllers\Admin\Edukasi\Webinar\WebinarController;
+use App\Http\Controllers\Creator\Kategori\CategoryController;
+use App\Http\Controllers\Creator\PortofolioController;
+use App\Http\Controllers\Creator\SkillController;
+use App\Http\Controllers\Creator\Pendidikan\PendidikanController;
+use App\Http\Controllers\FilterController;
+use App\Http\Controllers\LandingController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -14,9 +23,15 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', function () {
-    return view('welcome');
-});
+Route::get('/', [LandingController::class, 'home'])->name('home');
+Route::get('/landing/edukasi/video', [LandingController::class, 'video'])->name('video');
+
+Route::get('/landing/creator', [LandingController::class, 'creator'])->name('creator');
+Route::get('/landing/edukasi', [LandingController::class, 'edukasi'])->name('edukasi');
+Route::get('/landing/edukasi/webinar', [LandingController::class, 'webinar'])->name('webinar');
+Route::get('/landing/creator/detail/{id}', [LandingController::class, 'creatorDetail'])->name('creator.detail');
+Route::get('/landing/creator/filter', [FilterController::class, 'getCreatorByFilter'])->name('filter');
+
 
 Route::get('/dashboard', function () {
     return view('dashboard');
@@ -26,6 +41,46 @@ Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+
+    Route::get('/edukasi/login', [EdukasiController::class, 'index'])->name('edukasi.index');
+    Route::get('/edukasi/videos', [VideoController::class, 'index'])->name('videos.index');
+    Route::get('/edukasi/webinar', [WebinarController::class, 'index'])->name('webinar.index');
 });
 
-require __DIR__.'/auth.php';
+Route::middleware('creator')->group(function () {
+    Route::get('/creator/home', function () {
+        return view('creator.index');
+    })->name('creator.dashboard');
+    Route::resource('/creator/portofolio', PortofolioController::class);
+    Route::resource('/creator/skill', SkillController::class);
+
+    Route::get('/creator/pendidikan', [PendidikanController::class, 'index'])->name('pendidikan');
+    Route::get('/creator/pendidikan/create', [PendidikanController::class, 'create'])->name('pendidikan.create');
+    Route::post('/creator/pendidikan/create', [PendidikanController::class, 'store'])->name('pendidikan.store');
+    Route::get('/creator/pendidikan/{id}', [PendidikanController::class, 'edit'])->name('pendidikan.edit');
+    Route::put('/creator/pendidikan/update/{id}', [PendidikanController::class, 'update'])->name('pendidikan.update');
+    Route::get('/creator/pendidikan/delete/{id}', [PendidikanController::class, 'destroy'])->name('pendidikan.destroy');
+
+    Route::get('/creator/category', [CategoryController::class, 'index'])->name('kategori');
+    Route::post('/creator/category/create', [CategoryController::class, 'store'])->name('kategori.store');
+    Route::get('/creator/category/{id}', [CategoryController::class, 'edit'])->name('kategori.edit');
+    Route::put('/creator/category/update/{id}', [CategoryController::class, 'update'])->name('kategori.update');
+    Route::get('/creator/category/delete/{id}', [CategoryController::class, 'destroy'])->name('kategori.destroy');
+});
+
+Route::middleware('admin')->group(function () {
+    Route::get('/admin/edukasi/videos', [VideoController::class, 'create'])->name('videos.create');
+    Route::post('/admin/edukasi/videos', [VideoController::class, 'store'])->name('videos.store');
+    Route::get('/admin/edukasi/videos/{id}', [VideoController::class, 'edit'])->name('videos.edit');
+    Route::put('/admin/edukasi/videos/update/{id}', [VideoController::class, 'update'])->name('videos.update');
+    Route::get('/admin/edukasi/videos/delete/{id}', [VideoController::class, 'destroy'])->name('videos.destroy');
+
+    Route::get('/admin/edukasi/webinar', [WebinarController::class, 'create'])->name('webinar.create');
+    Route::post('/admin/edukasi/webinar', [WebinarController::class, 'store'])->name('webinar.store');
+    Route::get('/admin/edukasi/webinar/{id}', [WebinarController::class, 'edit'])->name('webinar.edit');
+    Route::put('/admin/edukasi/webinar/update/{id}', [WebinarController::class, 'update'])->name('webinar.update');
+    Route::get('/admin/edukasi/webinar/delete/{id}', [WebinarController::class, 'destroy'])->name('webinar.destroy');
+});
+
+
+require __DIR__ . '/auth.php';