+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # **Changelog**
 
+## [0.0.2] - 06-04-2023
+### Add
+- Add Laravel Breeze ([**@nichoadhyatma**](https://github.com/nichoadhyatma))
+- Add Migrations and Models ([**@nichoadhyatma**](https://github.com/nichoadhyatma))
+- Add Videos View ([**@maritsikmaly**](https://github.com/maritsikmaly))
+## Changed
+- 
 ## [0.0.1] - 06-04-2023
 🌱 initial release. ([**@maritsikmaly**](https://github.com/maritsikmaly))
-